+++ changelog.md
@@ -3,6 +3,7 @@
 # v3.1.6
 
 - **Feature**: You can now do `expiresAt: null` on entity update calls to remove the column (thus, stopping it from auto deleting)
+- **Feature**: `expiresAt` now accepts a `Date` as well
 - **Fix**: `expiresAt` not being accepted on `schema.from(entity).update({...})` - it would only show up on `entity.getUpdateParams`
 - **Fix**: `schema.from(entity).list({ filters })` was not correctly typed against the entity type. Now properly show inference and suggestions